requirements.txt
django==3.2.4
pillow==8.2.0
django-tinymce==3.3.0